+++ mathics/core/expression.py
@@ -2,20 +2,35 @@
 # -*- coding: utf-8 -*-
 
 import sympy
-import mpmath
 import math
-import re
 
 import typing
 from typing import Any, Optional
 from itertools import chain
 from bisect import bisect_left
-from functools import lru_cache
 
-
-from mathics.core.numbers import get_type, dps, prec, min_prec, machine_precision
+# from mathics.core.formatter import *
+from mathics.core.atoms import (
+    ByteArrayAtom,
+    Number,
+    Integer,
+    Rational,
+    Real,
+    Complex,
+    String,
+)
+from mathics.core.symbols import Atom, BaseExpression, Monomial, Symbol, system_symbols
+from mathics.core.systemsymbols import (
+    SymbolByteArray,
+    SymbolFalse,
+    SymbolList,
+    SymbolN,
+    SymbolNull,
+    SymbolSequence,
+    SymbolTrue,
+)
+from mathics.core.numbers import get_type, dps
 from mathics.core.convert import sympy_symbol_prefix, SympyExpression
-import base64
 
 # Imperical number that seems to work.
 # We have to be able to match mpmath values with sympy values
@@ -58,16 +73,6 @@ def strip_context(name) -> str:
     if "`" in name:
         return name[name.rindex("`") + 1 :]
     return name
-
-
-# system_symbols('A', 'B', ...) -> ['System`A', 'System`B', ...]
-def system_symbols(*symbols) -> typing.List[str]:
-    return [ensure_context(s) for s in symbols]
-
-
-# system_symbols_dict({'SomeSymbol': ...}) -> {'System`SomeSymbol': ...}
-def system_symbols_dict(d):
-    return {ensure_context(k): v for k, v in d.items()}
 
 
 class BoxError(Exception):
@@ -150,32 +155,32 @@ def from_python(arg):
         raise NotImplementedError
 
 
-class KeyComparable(object):
-    def get_sort_key(self):
-        raise NotImplementedError
+# class KeyComparable(object):
+#     def get_sort_key(self):
+#         raise NotImplementedError
 
-    def __lt__(self, other) -> bool:
-        return self.get_sort_key() < other.get_sort_key()
+#     def __lt__(self, other) -> bool:
+#         return self.get_sort_key() < other.get_sort_key()
 
-    def __gt__(self, other) -> bool:
-        return self.get_sort_key() > other.get_sort_key()
+#     def __gt__(self, other) -> bool:
+#         return self.get_sort_key() > other.get_sort_key()
 
-    def __le__(self, other) -> bool:
-        return self.get_sort_key() <= other.get_sort_key()
+#     def __le__(self, other) -> bool:
+#         return self.get_sort_key() <= other.get_sort_key()
 
-    def __ge__(self, other) -> bool:
-        return self.get_sort_key() >= other.get_sort_key()
+#     def __ge__(self, other) -> bool:
+#         return self.get_sort_key() >= other.get_sort_key()
 
-    def __eq__(self, other) -> bool:
-        return (
-            hasattr(other, "get_sort_key")
-            and self.get_sort_key() == other.get_sort_key()
-        )
+#     def __eq__(self, other) -> bool:
+#         return (
+#             hasattr(other, "get_sort_key")
+#             and self.get_sort_key() == other.get_sort_key()
+#         )
 
-    def __ne__(self, other) -> bool:
-        return (
-            not hasattr(other, "get_sort_key")
-        ) or self.get_sort_key() != other.get_sort_key()
+#     def __ne__(self, other) -> bool:
+#         return (
+#             not hasattr(other, "get_sort_key")
+#         ) or self.get_sort_key() != other.get_sort_key()
 
 
 # ExpressionCache keeps track of the following attributes for one Expression instance:
@@ -249,463 +254,6 @@ class ExpressionCache:
         return ExpressionCache(
             definitions.now, symbols, None if "System`Sequence" in symbols else tuple()
         )
-
-
-class BaseExpression(KeyComparable):
-    options: Any
-    pattern_sequence: bool
-    unformatted: Any
-    last_evaluated: Any
-
-    def __new__(cls, *args, **kwargs):
-        self = object.__new__(cls)
-        self.options = None
-        self.pattern_sequence = False
-        self.unformatted = self
-        self._cache = None
-        return self
-
-    def clear_cache(self):
-        self._cache = None
-
-    def equal2(self, rhs: Any) -> Optional[bool]:
-        """Mathics two-argument Equal (==)
-        returns True if self and rhs are identical.
-        """
-        if self.sameQ(rhs):
-            return True
-
-        # If the types are the same then we'll use the classes definition of == (or __eq__).
-        # Superclasses which need to specialized this behavior should redefine equal2()
-        #
-        # I would use `is` instead `==` here, to compare classes.
-        if type(self) is type(rhs):
-            return self == rhs
-        return None
-
-    def has_changed(self, definitions):
-        return True
-
-    def sequences(self):
-        return None
-
-    def flatten_sequence(self, evaluation) -> "BaseExpression":
-        return self
-
-    def flatten_pattern_sequence(self, evaluation) -> "BaseExpression":
-        return self
-
-    def get_attributes(self, definitions):
-        return set()
-
-    def evaluate_next(self, evaluation):
-        return self.evaluate(evaluation), False
-
-    def evaluate(self, evaluation) -> "BaseExpression":
-        evaluation.check_stopped()
-        return self
-
-    def get_atoms(self, include_heads=True):
-        return []
-
-    def get_name(self):
-        "Returns symbol's name if Symbol instance"
-
-        return ""
-
-    def is_symbol(self) -> bool:
-        return False
-
-    def is_machine_precision(self) -> bool:
-        return False
-
-    def get_lookup_name(self):
-        "Returns symbol name of leftmost head"
-
-        return self.get_name()
-
-    def get_head(self):
-        return None
-
-    def get_head_name(self):
-        return self.get_head().get_name()
-
-    def get_leaves(self):
-        return []
-
-    def get_int_value(self):
-        return None
-
-    def get_float_value(self, permit_complex=False):
-        return None
-
-    def get_string_value(self):
-        return None
-
-    def is_atom(self) -> bool:
-        return False
-
-    def is_true(self) -> bool:
-        return False
-
-    def is_numeric(self, evaluation=None) -> bool:
-        # used by NumericQ and expression ordering
-        return False
-
-    def has_form(self, heads, *leaf_counts):
-        return False
-
-    def flatten(self, head, pattern_only=False, callback=None) -> "BaseExpression":
-        return self
-
-    def __hash__(self):
-        """
-        To allow usage of expression as dictionary keys,
-        as in Expression.get_pre_choices
-        """
-        raise NotImplementedError
-
-    def user_hash(self, update) -> None:
-        # whereas __hash__ is for internal Mathics purposes like using Expressions as dictionary keys and fast
-        # comparison of elements, user_hash is called for Hash[]. user_hash should strive to give stable results
-        # across versions, whereas __hash__ must not. user_hash should try to hash all the data available, whereas
-        # __hash__ might only hash a sample of the data available.
-        raise NotImplementedError
-
-    def sameQ(self, rhs) -> bool:
-        """Mathics SameQ"""
-        return id(self) == id(rhs)
-
-    def get_sequence(self):
-        if self.get_head().get_name() == "System`Sequence":
-            return self.leaves
-        else:
-            return [self]
-
-    def evaluate_leaves(self, evaluation) -> "BaseExpression":
-        return self
-
-    def apply_rules(
-        self, rules, evaluation, level=0, options=None
-    ) -> typing.Tuple["BaseExpression", bool]:
-        if options:
-            l1, l2 = options["levelspec"]
-            if level < l1:
-                return self, False
-            elif l2 is not None and level > l2:
-                return self, False
-
-        for rule in rules:
-            result = rule.apply(self, evaluation, fully=False)
-            if result is not None:
-                return result, True
-        return self, False
-
-    def do_format(self, evaluation, form):
-        """
-        Applies formats associated to the expression and removes
-        superfluous enclosing formats.
-        """
-        formats = system_symbols(
-            "InputForm",
-            "OutputForm",
-            "StandardForm",
-            "FullForm",
-            "TraditionalForm",
-            "TeXForm",
-            "MathMLForm",
-        )
-
-        evaluation.inc_recursion_depth()
-        try:
-            expr = self
-            head = self.get_head_name()
-            leaves = self.get_leaves()
-            include_form = False
-            # If the expression is enclosed by a Format
-            # takes the form from the expression and
-            # removes the format from the expression.
-            if head in formats and len(leaves) == 1:
-                expr = leaves[0]
-                if not (form == "System`OutputForm" and head == "System`StandardForm"):
-                    form = head
-                    include_form = True
-            unformatted = expr
-            # If form is Fullform, return it without changes
-            if form == "System`FullForm":
-                if include_form:
-                    expr = Expression(form, expr)
-                    expr.unformatted = unformatted
-                return expr
-
-            # Repeated and RepeatedNull confuse the formatter,
-            # so we need to hardlink their format rules:
-            if head == "System`Repeated":
-                if len(leaves) == 1:
-                    return Expression(
-                        "System`HoldForm",
-                        Expression(
-                            "System`Postfix",
-                            Expression("System`List", leaves[0]),
-                            "..",
-                            170,
-                        ),
-                    )
-                else:
-                    return Expression("System`HoldForm", expr)
-            elif head == "System`RepeatedNull":
-                if len(leaves) == 1:
-                    return Expression(
-                        "System`HoldForm",
-                        Expression(
-                            "System`Postfix",
-                            Expression("System`List", leaves[0]),
-                            "...",
-                            170,
-                        ),
-                    )
-                else:
-                    return Expression("System`HoldForm", expr)
-
-            # If expr is not an atom, looks for formats in its definition
-            # and apply them.
-            def format_expr(expr):
-                if not (expr.is_atom()) and not (expr.head.is_atom()):
-                    # expr is of the form f[...][...]
-                    return None
-                name = expr.get_lookup_name()
-                formats = evaluation.definitions.get_formats(name, form)
-                for rule in formats:
-                    result = rule.apply(expr, evaluation)
-                    if result is not None and result != expr:
-                        return result.evaluate(evaluation)
-                return None
-
-            formatted = format_expr(expr)
-            if formatted is not None:
-                result = formatted.do_format(evaluation, form)
-                if include_form:
-                    result = Expression(form, result)
-                result.unformatted = unformatted
-                return result
-
-            # If the expression is still enclosed by a Format,
-            # iterate.
-            # If the expression is not atomic or of certain
-            # specific cases, iterate over the leaves.
-            head = expr.get_head_name()
-            if head in formats:
-                expr = expr.do_format(evaluation, form)
-            elif (
-                head != "System`NumberForm"
-                and not expr.is_atom()
-                and head != "System`Graphics"
-                and head != "System`Graphics3D"
-            ):
-                # print("Not inside graphics or numberform, and not is atom")
-                new_leaves = [leaf.do_format(evaluation, form) for leaf in expr.leaves]
-                expr = Expression(expr.head.do_format(evaluation, form), *new_leaves)
-
-            if include_form:
-                expr = Expression(form, expr)
-            expr.unformatted = unformatted
-            return expr
-        finally:
-            evaluation.dec_recursion_depth()
-
-    def format(
-        self, evaluation, form, **kwargs
-    ) -> typing.Union["Expression", "Symbol"]:
-        """
-        Applies formats associated to the expression, and then calls Makeboxes
-        """
-        expr = self.do_format(evaluation, form)
-        result = Expression("MakeBoxes", expr, Symbol(form)).evaluate(evaluation)
-        return result
-
-    def is_free(self, form, evaluation) -> bool:
-        from mathics.builtin.patterns import item_is_free
-
-        return item_is_free(self, form, evaluation)
-
-    def is_inexact(self) -> bool:
-        return self.get_precision() is not None
-
-    def get_precision(self):
-        return None
-
-    def get_option_values(self, evaluation, allow_symbols=False, stop_on_error=True):
-        options = self
-        if options.has_form("List", None):
-            options = options.flatten(SymbolList)
-            values = options.leaves
-        else:
-            values = [options]
-        option_values = {}
-        for option in values:
-            symbol_name = option.get_name()
-            if allow_symbols and symbol_name:
-                options = evaluation.definitions.get_options(symbol_name)
-                option_values.update(options)
-            else:
-                if not option.has_form(("Rule", "RuleDelayed"), 2):
-                    if stop_on_error:
-                        return None
-                    else:
-                        continue
-                name = option.leaves[0].get_name()
-                if not name and isinstance(option.leaves[0], String):
-                    name = ensure_context(option.leaves[0].get_string_value())
-                if not name:
-                    if stop_on_error:
-                        return None
-                    else:
-                        continue
-                option_values[name] = option.leaves[1]
-        return option_values
-
-    def get_rules_list(self):
-        from mathics.core.rules import Rule
-
-        list_expr = self.flatten(SymbolList)
-        list = []
-        if list_expr.has_form("List", None):
-            list.extend(list_expr.leaves)
-        else:
-            list.append(list_expr)
-        rules = []
-        for item in list:
-            if not item.has_form(("Rule", "RuleDelayed"), 2):
-                return None
-            rule = Rule(item.leaves[0], item.leaves[1])
-            rules.append(rule)
-        return rules
-
-    def to_sympy(self, **kwargs):
-        raise NotImplementedError
-
-    def to_mpmath(self):
-        return None
-
-    def round_to_float(self, evaluation=None, permit_complex=False):
-        """
-        Try to round to python float. Return None if not possible.
-        """
-        if evaluation is None:
-            value = self
-        elif isinstance(evaluation, sympy.core.numbers.NaN):
-            return None
-        else:
-            value = Expression(SymbolN, self).evaluate(evaluation)
-        if isinstance(value, Number):
-            value = value.round()
-            return value.get_float_value(permit_complex=permit_complex)
-
-    def __abs__(self) -> "Expression":
-        return Expression("Abs", self)
-
-    def __pos__(self):
-        return self
-
-    def __neg__(self):
-        return Expression("Times", self, -1)
-
-    def __add__(self, other) -> "Expression":
-        return Expression("Plus", self, other)
-
-    def __sub__(self, other) -> "Expression":
-        return Expression("Plus", self, Expression("Times", other, -1))
-
-    def __mul__(self, other) -> "Expression":
-        return Expression("Times", self, other)
-
-    def __truediv__(self, other) -> "Expression":
-        return Expression("Divide", self, other)
-
-    def __floordiv__(self, other) -> "Expression":
-        return Expression("Floor", Expression("Divide", self, other))
-
-    def __pow__(self, other) -> "Expression":
-        return Expression("Power", self, other)
-
-
-class Monomial(object):
-    """
-    An object to sort monomials, used in Expression.get_sort_key and
-    Symbol.get_sort_key.
-    """
-
-    def __init__(self, exps_dict):
-        self.exps = exps_dict
-
-    def __lt__(self, other) -> bool:
-        return self.__cmp(other) < 0
-
-    def __gt__(self, other) -> bool:
-        return self.__cmp(other) > 0
-
-    def __le__(self, other) -> bool:
-        return self.__cmp(other) <= 0
-
-    def __ge__(self, other) -> bool:
-        return self.__cmp(other) >= 0
-
-    def __eq__(self, other) -> bool:
-        return self.__cmp(other) == 0
-
-    def __ne__(self, other) -> bool:
-        return self.__cmp(other) != 0
-
-    def __cmp(self, other) -> int:
-        self_exps = self.exps.copy()
-        other_exps = other.exps.copy()
-        for var in self.exps:
-            if var in other.exps:
-                dec = min(self_exps[var], other_exps[var])
-                self_exps[var] -= dec
-                if not self_exps[var]:
-                    del self_exps[var]
-                other_exps[var] -= dec
-                if not other_exps[var]:
-                    del other_exps[var]
-        self_exps = sorted((var, exp) for var, exp in self_exps.items())
-        other_exps = sorted((var, exp) for var, exp in other_exps.items())
-
-        index = 0
-        self_len = len(self_exps)
-        other_len = len(other_exps)
-        while True:
-            if index >= self_len and index >= other_len:
-                return 0
-            if index >= self_len:
-                return -1  # self < other
-            if index >= other_len:
-                return 1  # self > other
-            self_var, self_exp = self_exps[index]
-            other_var, other_exp = other_exps[index]
-            if self_var < other_var:
-                return -1
-            if self_var > other_var:
-                return 1
-            if self_exp != other_exp:
-                if index + 1 == self_len or index + 1 == other_len:
-                    # smaller exponents first
-                    if self_exp < other_exp:
-                        return -1
-                    elif self_exp == other_exp:
-                        return 0
-                    else:
-                        return 1
-                else:
-                    # bigger exponents first
-                    if self_exp < other_exp:
-                        return 1
-                    elif self_exp == other_exp:
-                        return 0
-                    else:
-                        return -1
-            index += 1
-        return 0
 
 
 class Expression(BaseExpression):
@@ -1839,7 +1387,7 @@ class Expression(BaseExpression):
 
     def is_numeric(self, evaluation=None) -> bool:
         if evaluation:
-            if not "System`NumericFunction" in evaluation.definitions.get_attributes(
+            if "System`NumericFunction" not in evaluation.definitions.get_attributes(
                 self._head.get_name()
             ):
                 return False
@@ -1903,1197 +1451,6 @@ class Expression(BaseExpression):
 
     def __getnewargs__(self):
         return (self._head, self._leaves)
-
-
-class Atom(BaseExpression):
-    def is_atom(self) -> bool:
-        return True
-
-    def equal2(self, rhs: Any) -> Optional[bool]:
-        """Mathics two-argument Equal (==)
-        returns True if self and rhs are identical.
-        """
-        if self.sameQ(rhs):
-            return True
-        if isinstance(rhs, Symbol) or not isinstance(rhs, Atom):
-            return None
-        return self == rhs
-
-    def has_form(self, heads, *leaf_counts) -> bool:
-        if leaf_counts:
-            return False
-        name = self.get_atom_name()
-        if isinstance(heads, tuple):
-            return name in heads
-        else:
-            return heads == name
-
-    def has_symbol(self, symbol_name) -> bool:
-        return False
-
-    def get_head(self) -> "Symbol":
-        return Symbol(self.get_atom_name())
-
-    def get_atom_name(self) -> str:
-        return self.__class__.__name__
-
-    def __repr__(self) -> str:
-        return "<%s: %s>" % (self.get_atom_name(), self)
-
-    def replace_vars(self, vars, options=None, in_scoping=True) -> "Atom":
-        return self
-
-    def replace_slots(self, slots, evaluation) -> "Atom":
-        return self
-
-    def numerify(self, evaluation) -> "Atom":
-        return self
-
-    def copy(self, reevaluate=False) -> "Atom":
-        result = self.do_copy()
-        result.original = self
-        return result
-
-    def set_positions(self, position=None) -> None:
-        self.position = position
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return [0, 0, 1, 1, 0, 0, 0, 1]
-        else:
-            raise NotImplementedError
-
-    def get_atoms(self, include_heads=True) -> typing.List["Atom"]:
-        return [self]
-
-    def atom_to_boxes(self, f, evaluation):
-        raise NotImplementedError
-
-
-class Symbol(Atom):
-    name: str
-    sympy_dummy: Any
-    defined_symbols = {}
-
-    def __new__(cls, name, sympy_dummy=None):
-        name = ensure_context(name)
-        self = cls.defined_symbols.get(name, None)
-        if self is None:
-            self = super(Symbol, cls).__new__(cls)
-            self.name = name
-            self.sympy_dummy = sympy_dummy
-            # cls.defined_symbols[name] = self
-        return self
-
-    def __str__(self) -> str:
-        return self.name
-
-    def do_copy(self) -> "Symbol":
-        return Symbol(self.name)
-
-    def boxes_to_text(self, **options) -> str:
-        return str(self.name)
-
-    def atom_to_boxes(self, f, evaluation) -> "String":
-        return String(evaluation.definitions.shorten_name(self.name))
-
-    def to_sympy(self, **kwargs):
-        from mathics.builtin import mathics_to_sympy
-
-        if self.sympy_dummy is not None:
-            return self.sympy_dummy
-
-        builtin = mathics_to_sympy.get(self.name)
-        if (
-            builtin is None
-            or not builtin.sympy_name
-            or not builtin.is_constant()  # nopep8
-        ):
-            return sympy.Symbol(sympy_symbol_prefix + self.name)
-        return builtin.to_sympy(self, **kwargs)
-
-    def to_python(self, *args, **kwargs):
-        if self == SymbolTrue:
-            return True
-        if self == SymbolFalse:
-            return False
-        if self == SymbolNull:
-            return None
-        n_evaluation = kwargs.get("n_evaluation")
-        if n_evaluation is not None:
-            value = Expression(SymbolN, self).evaluate(n_evaluation)
-            return value.to_python()
-
-        if kwargs.get("python_form", False):
-            return self.to_sympy(**kwargs)
-        else:
-            return self.name
-
-    def default_format(self, evaluation, form) -> str:
-        return self.name
-
-    def get_attributes(self, definitions):
-        return definitions.get_attributes(self.name)
-
-    def get_name(self) -> str:
-        return self.name
-
-    def is_symbol(self) -> bool:
-        return True
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super(Symbol, self).get_sort_key(True)
-        else:
-            return [
-                1 if self.is_numeric() else 2,
-                2,
-                Monomial({self.name: 1}),
-                0,
-                self.name,
-                1,
-            ]
-
-    def equal2(self, rhs: Any) -> Optional[bool]:
-        """Mathics two-argument Equal (==)"""
-        if self.sameQ(rhs):
-            return True
-
-        # Booleans are treated like constants, but all other symbols
-        # are treated None. We could create a Bool class and
-        # define equal2 in that, but for just this doesn't
-        # seem to be worth it. If other things come up, this may change.
-        if self in (SymbolTrue, SymbolFalse) and rhs in (SymbolTrue, SymbolFalse):
-            return self == rhs
-        return None
-
-    def sameQ(self, rhs: Any) -> bool:
-        """Mathics SameQ"""
-        return id(self) == id(rhs) or isinstance(rhs, Symbol) and self.name == rhs.name
-
-    def replace_vars(self, vars, options={}, in_scoping=True):
-        assert all(fully_qualified_symbol_name(v) for v in vars)
-        var = vars.get(self.name, None)
-        if var is None:
-            return self
-        else:
-            return var
-
-    def has_symbol(self, symbol_name) -> bool:
-        return self.name == ensure_context(symbol_name)
-
-    def evaluate(self, evaluation):
-        rules = evaluation.definitions.get_ownvalues(self.name)
-        for rule in rules:
-            result = rule.apply(self, evaluation, fully=True)
-            if result is not None and not result.sameQ(self):
-                return result.evaluate(evaluation)
-        return self
-
-    def is_true(self) -> bool:
-        return self == SymbolTrue
-
-    def is_numeric(self, evaluation=None) -> bool:
-        return self.name in system_symbols(
-            "Pi", "E", "EulerGamma", "GoldenRatio", "MachinePrecision", "Catalan"
-        )
-
-    def __hash__(self):
-        return hash(("Symbol", self.name))  # to distinguish from String
-
-    def user_hash(self, update) -> None:
-        update(b"System`Symbol>" + self.name.encode("utf8"))
-
-    def __getnewargs__(self):
-        return (self.name, self.sympy_dummy)
-
-
-# Some common Symbols. This list is sorted in alpabetic order.
-SymbolAborted = Symbol("$Aborted")
-SymbolAssociation = Symbol("Association")
-SymbolByteArray = Symbol("ByteArray")
-SymbolComplexInfinity = Symbol("ComplexInfinity")
-SymbolDirectedInfinity = Symbol("DirectedInfinity")
-SymbolFailed = Symbol("$Failed")
-SymbolFalse = Symbol("False")
-SymbolInfinity = Symbol("Infinity")
-SymbolList = Symbol("List")
-SymbolMachinePrecision = Symbol("MachinePrecision")
-SymbolMakeBoxes = Symbol("MakeBoxes")
-SymbolN = Symbol("N")
-SymbolNull = Symbol("Null")
-SymbolRule = Symbol("Rule")
-SymbolSequence = Symbol("Sequence")
-SymbolTrue = Symbol("True")
-SymbolUndefined = Symbol("Undefined")
-SymbolLess = Symbol("Less")
-SymbolGreater = Symbol("Greater")
-
-
-@lru_cache(maxsize=1024)
-def from_mpmath(value, prec=None):
-    "Converts mpf or mpc to Number."
-    if isinstance(value, mpmath.mpf):
-        if prec is None:
-            return MachineReal(float(value))
-        else:
-            # HACK: use str here to prevent loss of precision
-            return PrecisionReal(sympy.Float(str(value), prec))
-    elif isinstance(value, mpmath.mpc):
-        if value.imag == 0.0:
-            return from_mpmath(value.real, prec)
-        real = from_mpmath(value.real, prec)
-        imag = from_mpmath(value.imag, prec)
-        return Complex(real, imag)
-    else:
-        raise TypeError(type(value))
-
-
-class Number(Atom):
-    def __str__(self) -> str:
-        return str(self.value)
-
-    def is_numeric(self, evaluation=None) -> bool:
-        return True
-
-
-def _ExponentFunction(value):
-    n = value.get_int_value()
-    if -5 <= n <= 5:
-        return SymbolNull
-    else:
-        return value
-
-
-def _NumberFormat(man, base, exp, options):
-    if exp.get_string_value():
-        if options["_Form"] in (
-            "System`InputForm",
-            "System`OutputForm",
-            "System`FullForm",
-        ):
-            return Expression("RowBox", Expression(SymbolList, man, String("*^"), exp))
-        else:
-            return Expression(
-                "RowBox",
-                Expression(
-                    "List",
-                    man,
-                    String(options["NumberMultiplier"]),
-                    Expression("SuperscriptBox", base, exp),
-                ),
-            )
-    else:
-        return man
-
-
-_number_form_options = {
-    "DigitBlock": [0, 0],
-    "ExponentFunction": _ExponentFunction,
-    "ExponentStep": 1,
-    "NumberFormat": _NumberFormat,
-    "NumberPadding": ["", "0"],
-    "NumberPoint": ".",
-    "NumberSigns": ["-", ""],
-    "SignPadding": False,
-    "NumberMultiplier": "\u00d7",
-}
-
-
-class Integer(Number):
-    value: int
-
-    def __new__(cls, value) -> "Integer":
-        n = int(value)
-        self = super(Integer, cls).__new__(cls)
-        self.value = n
-        return self
-
-    @lru_cache()
-    def __init__(self, value) -> "Integer":
-        super().__init__()
-
-    def boxes_to_text(self, **options) -> str:
-        return str(self.value)
-
-    def boxes_to_mathml(self, **options) -> str:
-        return self.make_boxes("MathMLForm").boxes_to_mathml(**options)
-
-    def boxes_to_tex(self, **options) -> str:
-        return str(self.value)
-
-    def make_boxes(self, form) -> "String":
-        return String(str(self.value))
-
-    def atom_to_boxes(self, f, evaluation):
-        return self.make_boxes(f.get_name())
-
-    def default_format(self, evaluation, form) -> str:
-        return str(self.value)
-
-    def to_sympy(self, **kwargs):
-        return sympy.Integer(self.value)
-
-    def to_mpmath(self):
-        return mpmath.mpf(self.value)
-
-    def to_python(self, *args, **kwargs):
-        return self.value
-
-    def round(self, d=None) -> typing.Union["MachineReal", "PrecisionReal"]:
-        if d is None:
-            return MachineReal(float(self.value))
-        else:
-            return PrecisionReal(sympy.Float(self.value, d))
-
-    def get_int_value(self) -> int:
-        return self.value
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        return isinstance(other, Integer) and self.value == other.value
-
-    def evaluate(self, evaluation):
-        evaluation.check_stopped()
-        return self
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        else:
-            return [0, 0, self.value, 0, 1]
-
-    def do_copy(self) -> "Integer":
-        return Integer(self.value)
-
-    def __hash__(self):
-        return hash(("Integer", self.value))
-
-    def user_hash(self, update):
-        update(b"System`Integer>" + str(self.value).encode("utf8"))
-
-    def __getnewargs__(self):
-        return (self.value,)
-
-    def __neg__(self) -> "Integer":
-        return Integer(-self.value)
-
-    @property
-    def is_zero(self) -> bool:
-        return self.value == 0
-
-
-Integer0 = Integer(0)
-Integer1 = Integer(1)
-
-
-class Rational(Number):
-    @lru_cache()
-    def __new__(cls, numerator, denominator=1) -> "Rational":
-        self = super().__new__(cls)
-        self.value = sympy.Rational(numerator, denominator)
-        return self
-
-    def atom_to_boxes(self, f, evaluation):
-        return self.format(evaluation, f.get_name())
-
-    def to_sympy(self, **kwargs):
-        return self.value
-
-    def to_mpmath(self):
-        return mpmath.mpf(self.value)
-
-    def to_python(self, *args, **kwargs) -> float:
-        return float(self.value)
-
-    def round(self, d=None) -> typing.Union["MachineReal", "PrecisionReal"]:
-        if d is None:
-            return MachineReal(float(self.value))
-        else:
-            return PrecisionReal(self.value.n(d))
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        return isinstance(other, Rational) and self.value == other.value
-
-    def numerator(self) -> "Integer":
-        return Integer(self.value.as_numer_denom()[0])
-
-    def denominator(self) -> "Integer":
-        return Integer(self.value.as_numer_denom()[1])
-
-    def do_format(self, evaluation, form) -> "Expression":
-        assert fully_qualified_symbol_name(form)
-        if form == "System`FullForm":
-            return Expression(
-                Expression("HoldForm", Symbol("Rational")),
-                self.numerator(),
-                self.denominator(),
-            ).do_format(evaluation, form)
-        else:
-            numerator = self.numerator()
-            minus = numerator.value < 0
-            if minus:
-                numerator = Integer(-numerator.value)
-            result = Expression("Divide", numerator, self.denominator())
-            if minus:
-                result = Expression("Minus", result)
-            result = Expression("HoldForm", result)
-            return result.do_format(evaluation, form)
-
-    def default_format(self, evaluation, form) -> str:
-        return "Rational[%s, %s]" % self.value.as_numer_denom()
-
-    def evaluate(self, evaluation) -> "Rational":
-        evaluation.check_stopped()
-        return self
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        else:
-            # HACK: otherwise "Bus error" when comparing 1==1.
-            return [0, 0, sympy.Float(self.value), 0, 1]
-
-    def do_copy(self) -> "Rational":
-        return Rational(self.value)
-
-    def __hash__(self):
-        return hash(("Rational", self.value))
-
-    def user_hash(self, update) -> None:
-        update(
-            b"System`Rational>" + ("%s>%s" % self.value.as_numer_denom()).encode("utf8")
-        )
-
-    def __getnewargs__(self):
-        return (self.numerator().get_int_value(), self.denominator().get_int_value())
-
-    def __neg__(self) -> "Rational":
-        return Rational(
-            -self.numerator().get_int_value(), self.denominator().get_int_value()
-        )
-
-    @property
-    def is_zero(self) -> bool:
-        return (
-            self.numerator().is_zero
-        )  # (implicit) and not (self.denominator().is_zero)
-
-
-RationalOneHalf = Rational(1, 2)
-
-
-class Real(Number):
-    def __new__(cls, value, p=None) -> "Real":
-        if isinstance(value, str):
-            value = str(value)
-            if p is None:
-                digits = ("".join(re.findall("[0-9]+", value))).lstrip("0")
-                if digits == "":  # Handle weird Mathematica zero case
-                    p = max(prec(len(value.replace("0.", ""))), machine_precision)
-                else:
-                    p = prec(len(digits.zfill(dps(machine_precision))))
-        elif isinstance(value, sympy.Float):
-            if p is None:
-                p = value._prec + 1
-        elif isinstance(value, (Integer, sympy.Number, mpmath.mpf, float, int)):
-            if p is not None and p > machine_precision:
-                value = str(value)
-        else:
-            raise TypeError("Unknown number type: %s (type %s)" % (value, type(value)))
-
-        # return either machine precision or arbitrary precision real
-        if p is None or p == machine_precision:
-            return MachineReal.__new__(MachineReal, value)
-        else:
-            return PrecisionReal.__new__(PrecisionReal, value)
-
-    def boxes_to_text(self, **options) -> str:
-        return self.make_boxes("System`OutputForm").boxes_to_text(**options)
-
-    def boxes_to_mathml(self, **options) -> str:
-        return self.make_boxes("System`MathMLForm").boxes_to_mathml(**options)
-
-    def boxes_to_tex(self, **options) -> str:
-        return self.make_boxes("System`TeXForm").boxes_to_tex(**options)
-
-    def atom_to_boxes(self, f, evaluation):
-        return self.make_boxes(f.get_name())
-
-    def evaluate(self, evaluation) -> "Real":
-        evaluation.check_stopped()
-        return self
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        return [0, 0, self.value, 0, 1]
-
-    def is_nan(self, d=None) -> bool:
-        return isinstance(self.value, sympy.core.numbers.NaN)
-
-    def __eq__(self, other) -> bool:
-        if isinstance(other, Real):
-            # MMA Docs: "Approximate numbers that differ in their last seven
-            # binary digits are considered equal"
-            _prec = min_prec(self, other)
-            with mpmath.workprec(_prec):
-                rel_eps = 0.5 ** (_prec - 7)
-                return mpmath.almosteq(
-                    self.to_mpmath(), other.to_mpmath(), abs_eps=0, rel_eps=rel_eps
-                )
-        else:
-            return self.get_sort_key() == other.get_sort_key()
-
-    def __ne__(self, other) -> bool:
-        # Real is a total order
-        return not (self == other)
-
-    def __hash__(self):
-        # ignore last 7 binary digits when hashing
-        _prec = self.get_precision()
-        return hash(("Real", self.to_sympy().n(dps(_prec))))
-
-    def user_hash(self, update):
-        # ignore last 7 binary digits when hashing
-        _prec = self.get_precision()
-        update(b"System`Real>" + str(self.to_sympy().n(dps(_prec))).encode("utf8"))
-
-    def get_atom_name(self) -> str:
-        return "Real"
-
-
-class MachineReal(Real):
-    """
-    Machine precision real number.
-
-    Stored internally as a python float.
-    """
-
-    value: float
-
-    def __new__(cls, value) -> "MachineReal":
-        self = Number.__new__(cls)
-        self.value = float(value)
-        if math.isinf(self.value) or math.isnan(self.value):
-            raise OverflowError
-        return self
-
-    def to_python(self, *args, **kwargs) -> float:
-        return self.value
-
-    def to_sympy(self, *args, **kwargs):
-        return sympy.Float(self.value)
-
-    def to_mpmath(self):
-        return mpmath.mpf(self.value)
-
-    def round(self, d=None) -> "MachineReal":
-        return self
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        if isinstance(other, MachineReal):
-            return self.value == other.value
-        elif isinstance(other, PrecisionReal):
-            return self.to_sympy() == other.value
-        return False
-
-    def is_machine_precision(self) -> bool:
-        return True
-
-    def get_precision(self) -> int:
-        return machine_precision
-
-    def get_float_value(self, permit_complex=False) -> float:
-        return self.value
-
-    def make_boxes(self, form):
-        from mathics.builtin.inout import number_form
-
-        _number_form_options["_Form"] = form  # passed to _NumberFormat
-        if form in ("System`InputForm", "System`FullForm"):
-            n = None
-        else:
-            n = 6
-        return number_form(self, n, None, None, _number_form_options)
-
-    def __getnewargs__(self):
-        return (self.value,)
-
-    def do_copy(self) -> "MachineReal":
-        return MachineReal(self.value)
-
-    def __neg__(self) -> "MachineReal":
-        return MachineReal(-self.value)
-
-    @property
-    def is_zero(self) -> bool:
-        return self.value == 0.0
-
-    @property
-    def is_approx_zero(self) -> bool:
-        # In WMA, Chop[10.^(-10)] == 0,
-        # so, lets take it.
-        res = abs(self.value) <= 1e-10
-        return res
-
-
-class PrecisionReal(Real):
-    """
-    Arbitrary precision real number.
-
-    Stored internally as a sympy.Float.
-
-    Note: Plays nicely with the mpmath.mpf (float) type.
-    """
-
-    value: sympy.Float
-
-    def __new__(cls, value) -> "PrecisionReal":
-        self = Number.__new__(cls)
-        self.value = sympy.Float(value)
-        return self
-
-    def to_python(self, *args, **kwargs):
-        return float(self.value)
-
-    def to_sympy(self, *args, **kwargs):
-        return self.value
-
-    def to_mpmath(self):
-        return mpmath.mpf(self.value)
-
-    def round(self, d=None) -> typing.Union["MachineReal", "PrecisionReal"]:
-        if d is None:
-            return MachineReal(float(self.value))
-        else:
-            d = min(dps(self.get_precision()), d)
-            return PrecisionReal(self.value.n(d))
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        if isinstance(other, PrecisionReal):
-            return self.value == other.value
-        elif isinstance(other, MachineReal):
-            return self.value == other.to_sympy()
-        return False
-
-    def get_precision(self) -> int:
-        return self.value._prec + 1
-
-    def make_boxes(self, form):
-        from mathics.builtin.inout import number_form
-
-        _number_form_options["_Form"] = form  # passed to _NumberFormat
-        return number_form(
-            self, dps(self.get_precision()), None, None, _number_form_options
-        )
-
-    def __getnewargs__(self):
-        return (self.value,)
-
-    def do_copy(self) -> "PrecisionReal":
-        return PrecisionReal(self.value)
-
-    def __neg__(self) -> "PrecisionReal":
-        return PrecisionReal(-self.value)
-
-    @property
-    def is_zero(self) -> bool:
-        return self.value == 0.0
-
-
-class Complex(Number):
-    """
-    Complex wraps two real-valued Numbers.
-    """
-
-    real: Any
-    imag: Any
-
-    def __new__(cls, real, imag):
-        self = super().__new__(cls)
-        if isinstance(real, Complex) or not isinstance(real, Number):
-            raise ValueError("Argument 'real' must be a real number.")
-        if isinstance(imag, Complex) or not isinstance(imag, Number):
-            raise ValueError("Argument 'imag' must be a real number.")
-
-        if imag.sameQ(Integer0):
-            return real
-
-        if isinstance(real, MachineReal) and not isinstance(imag, MachineReal):
-            imag = imag.round()
-        if isinstance(imag, MachineReal) and not isinstance(real, MachineReal):
-            real = real.round()
-
-        self.real = real
-        self.imag = imag
-        return self
-
-    def atom_to_boxes(self, f, evaluation):
-        return self.format(evaluation, f.get_name())
-
-    def __str__(self) -> str:
-        return str(self.to_sympy())
-
-    def to_sympy(self, **kwargs):
-        return self.real.to_sympy() + sympy.I * self.imag.to_sympy()
-
-    def to_python(self, *args, **kwargs):
-        return complex(
-            self.real.to_python(*args, **kwargs), self.imag.to_python(*args, **kwargs)
-        )
-
-    def to_mpmath(self):
-        return mpmath.mpc(self.real.to_mpmath(), self.imag.to_mpmath())
-
-    def do_format(self, evaluation, form) -> "Expression":
-        if form == "System`FullForm":
-            return Expression(
-                Expression("HoldForm", Symbol("Complex")), self.real, self.imag
-            ).do_format(evaluation, form)
-
-        parts: typing.List[Any] = []
-        if self.is_machine_precision() or not self.real.is_zero:
-            parts.append(self.real)
-        if self.imag.sameQ(Integer(1)):
-            parts.append(Symbol("I"))
-        else:
-            parts.append(Expression("Times", self.imag, Symbol("I")))
-
-        if len(parts) == 1:
-            result = parts[0]
-        else:
-            result = Expression("Plus", *parts)
-
-        return Expression("HoldForm", result).do_format(evaluation, form)
-
-    def default_format(self, evaluation, form) -> str:
-        return "Complex[%s, %s]" % (
-            self.real.default_format(evaluation, form),
-            self.imag.default_format(evaluation, form),
-        )
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        else:
-            return [0, 0, self.real.get_sort_key()[2], self.imag.get_sort_key()[2], 1]
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        return (
-            isinstance(other, Complex)
-            and self.real == other.real
-            and self.imag == other.imag
-        )
-
-    def evaluate(self, evaluation) -> "Complex":
-        evaluation.check_stopped()
-        return self
-
-    def round(self, d=None) -> "Complex":
-        real = self.real.round(d)
-        imag = self.imag.round(d)
-        return Complex(real, imag)
-
-    def is_machine_precision(self) -> bool:
-        if self.real.is_machine_precision() or self.imag.is_machine_precision():
-            return True
-        return False
-
-    def get_float_value(self, permit_complex=False) -> typing.Optional[complex]:
-        if permit_complex:
-            real = self.real.get_float_value()
-            imag = self.imag.get_float_value()
-            if real is not None and imag is not None:
-                return complex(real, imag)
-        else:
-            return None
-
-    def get_precision(self) -> typing.Optional[int]:
-        real_prec = self.real.get_precision()
-        imag_prec = self.imag.get_precision()
-        if imag_prec is None or real_prec is None:
-            return None
-        return min(real_prec, imag_prec)
-
-    def do_copy(self) -> "Complex":
-        return Complex(self.real.do_copy(), self.imag.do_copy())
-
-    def __hash__(self):
-        return hash(("Complex", self.real, self.imag))
-
-    def user_hash(self, update) -> None:
-        update(b"System`Complex>")
-        update(self.real)
-        update(self.imag)
-
-    def __eq__(self, other) -> bool:
-        if isinstance(other, Complex):
-            return self.real == other.real and self.imag == other.imag
-        else:
-            return self.get_sort_key() == other.get_sort_key()
-
-    def __getnewargs__(self):
-        return (self.real, self.imag)
-
-    def __neg__(self):
-        return Complex(-self.real, -self.imag)
-
-    @property
-    def is_zero(self) -> bool:
-        return self.real.is_zero and self.imag.is_zero
-
-    @property
-    def is_approx_zero(self) -> bool:
-        real_zero = (
-            self.real.is_approx_zero
-            if hasattr(self.real, "is_approx_zero")
-            else self.real.is_zero
-        )
-        imag_zero = (
-            self.imag.is_approx_zero
-            if hasattr(self.imag, "is_approx_zero")
-            else self.imag.is_zero
-        )
-        return real_zero and imag_zero
-
-
-def encode_mathml(text: str) -> str:
-    text = text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;")
-    text = text.replace('"', "&quot;").replace(" ", "&nbsp;")
-    text = text.replace("\n", '<mspace linebreak="newline" />')
-    return text
-
-
-TEX_REPLACE = {
-    "{": r"\{",
-    "}": r"\}",
-    "_": r"\_",
-    "$": r"\$",
-    "%": r"\%",
-    "#": r"\#",
-    "&": r"\&",
-    "\\": r"\backslash{}",
-    "^": r"{}^{\wedge}",
-    "~": r"\sim{}",
-    "|": r"\vert{}",
-}
-TEX_TEXT_REPLACE = TEX_REPLACE.copy()
-TEX_TEXT_REPLACE.update(
-    {
-        "<": r"$<$",
-        ">": r"$>$",
-        "~": r"$\sim$",
-        "|": r"$\vert$",
-        "\\": r"$\backslash$",
-        "^": r"${}^{\wedge}$",
-    }
-)
-TEX_REPLACE_RE = re.compile("([" + "".join([re.escape(c) for c in TEX_REPLACE]) + "])")
-
-
-def encode_tex(text: str, in_text=False) -> str:
-    def replace(match):
-        c = match.group(1)
-        repl = TEX_TEXT_REPLACE if in_text else TEX_REPLACE
-        # return TEX_REPLACE[c]
-        return repl.get(c, c)
-
-    text = TEX_REPLACE_RE.sub(replace, text)
-    text = text.replace("\n", "\\newline\n")
-    return text
-
-
-extra_operators = set(
-    (
-        ",",
-        "(",
-        ")",
-        "[",
-        "]",
-        "{",
-        "}",
-        "\u301a",
-        "\u301b",
-        "\u00d7",
-        "\u2032",
-        "\u2032\u2032",
-        " ",
-        "\u2062",
-        "\u222b",
-        "\u2146",
-    )
-)
-
-
-class String(Atom):
-    value: str
-
-    def __new__(cls, value):
-        self = super().__new__(cls)
-        self.value = str(value)
-        return self
-
-    def __str__(self) -> str:
-        return '"%s"' % self.value
-
-    def boxes_to_text(self, show_string_characters=False, **options) -> str:
-        value = self.value
-
-        if (
-            not show_string_characters
-            and value.startswith('"')  # nopep8
-            and value.endswith('"')
-        ):
-            value = value[1:-1]
-
-        return value
-
-    def boxes_to_mathml(self, show_string_characters=False, **options) -> str:
-        from mathics.core.parser import is_symbol_name
-        from mathics.builtin import builtins_by_module
-
-        operators = set()
-        for modname, builtins in builtins_by_module.items():
-            for builtin in builtins:
-                # name = builtin.get_name()
-                operator = builtin.get_operator_display()
-                if operator is not None:
-                    operators.add(operator)
-
-        text = self.value
-
-        def render(format, string):
-            encoded_text = encode_mathml(string)
-            return format % encoded_text
-
-        if text.startswith('"') and text.endswith('"'):
-            if show_string_characters:
-                return render("<ms>%s</ms>", text[1:-1])
-            else:
-                outtext = ""
-                for line in text[1:-1].split("\n"):
-                    outtext += render("<mtext>%s</mtext>", line)
-                return outtext
-        elif text and ("0" <= text[0] <= "9" or text[0] == "."):
-            return render("<mn>%s</mn>", text)
-        else:
-            if text in operators or text in extra_operators:
-                if text == "\u2146":
-                    return render(
-                        '<mo form="prefix" lspace="0.2em" rspace="0">%s</mo>', text
-                    )
-                if text == "\u2062":
-                    return render(
-                        '<mo form="prefix" lspace="0" rspace="0.2em">%s</mo>', text
-                    )
-                return render("<mo>%s</mo>", text)
-            elif is_symbol_name(text):
-                return render("<mi>%s</mi>", text)
-            else:
-                outtext = ""
-                for line in text.split("\n"):
-                    outtext += render("<mtext>%s</mtext>", line)
-                return outtext
-
-    def boxes_to_tex(self, show_string_characters=False, **options) -> str:
-        from mathics.builtin import builtins_by_module
-
-        operators = set()
-
-        for modname, builtins in builtins_by_module.items():
-            for builtin in builtins:
-                operator = builtin.get_operator_display()
-                if operator is not None:
-                    operators.add(operator)
-
-        text = self.value
-
-        def render(format, string, in_text=False):
-            return format % encode_tex(string, in_text)
-
-        if text.startswith('"') and text.endswith('"'):
-            if show_string_characters:
-                return render(r'\text{"%s"}', text[1:-1], in_text=True)
-            else:
-                return render(r"\text{%s}", text[1:-1], in_text=True)
-        elif text and text[0] in "0123456789-.":
-            return render("%s", text)
-        else:
-            # FIXME: this should be done in a better way.
-            if text == "\u2032":
-                return "'"
-            elif text == "\u2032\u2032":
-                return "''"
-            elif text == "\u2062":
-                return " "
-            elif text == "\u221e":
-                return r"\infty "
-            elif text == "\u00d7":
-                return r"\times "
-            elif text in ("(", "[", "{"):
-                return render(r"\left%s", text)
-            elif text in (")", "]", "}"):
-                return render(r"\right%s", text)
-            elif text == "\u301a":
-                return r"\left[\left["
-            elif text == "\u301b":
-                return r"\right]\right]"
-            elif text == "," or text == ", ":
-                return text
-            elif text == "\u222b":
-                return r"\int"
-            # Tolerate WL or Unicode DifferentialD
-            elif text in ("\u2146", "\U0001D451"):
-                return r"\, d"
-            elif text == "\u2211":
-                return r"\sum"
-            elif text == "\u220f":
-                return r"\prod"
-            elif len(text) > 1:
-                return render(r"\text{%s}", text, in_text=True)
-            else:
-                return render("%s", text)
-
-    def atom_to_boxes(self, f, evaluation):
-        inner = str(self.value)
-
-        if f.get_name() in system_symbols("InputForm", "FullForm"):
-            inner = inner.replace("\\", "\\\\")
-
-        return String('"' + inner + '"')
-
-    def do_copy(self) -> "String":
-        return String(self.value)
-
-    def default_format(self, evaluation, form) -> str:
-        value = self.value.replace("\\", "\\\\").replace('"', '\\"')
-        return '"%s"' % value
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        else:
-            return [0, 1, self.value, 0, 1]
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        return isinstance(other, String) and self.value == other.value
-
-    def get_string_value(self) -> str:
-        return self.value
-
-    def to_sympy(self, **kwargs):
-        return None
-
-    def to_python(self, *args, **kwargs) -> str:
-        if kwargs.get("string_quotes", True):
-            return '"%s"' % self.value  # add quotes to distinguish from Symbols
-        else:
-            return self.value
-
-    def __hash__(self):
-        return hash(("String", self.value))
-
-    def user_hash(self, update):
-        # hashing a String is the one case where the user gets the untampered
-        # hash value of the string's text. this corresponds to MMA behavior.
-        update(self.value.encode("utf8"))
-
-    def __getnewargs__(self):
-        return (self.value,)
-
-
-class ByteArrayAtom(Atom):
-    value: str
-
-    def __new__(cls, value):
-        self = super().__new__(cls)
-        if type(value) in (bytes, bytearray):
-            self.value = value
-        elif type(value) is list:
-            self.value = bytearray(list)
-        elif type(value) is str:
-            self.value = base64.b64decode(value)
-        else:
-            raise Exception("value does not belongs to a valid type")
-        return self
-
-    def __str__(self) -> str:
-        return base64.b64encode(self.value).decode("utf8")
-
-    def boxes_to_text(self, **options) -> str:
-        return '"' + self.__str__() + '"'
-
-    def boxes_to_mathml(self, **options) -> str:
-        return encode_mathml(String('"' + self.__str__() + '"'))
-
-    def boxes_to_tex(self, **options) -> str:
-        return encode_tex(String('"' + self.__str__() + '"'))
-
-    def atom_to_boxes(self, f, evaluation):
-        res = String('""' + self.__str__() + '""')
-        return res
-
-    def do_copy(self) -> "ByteArrayAtom":
-        return ByteArrayAtom(self.value)
-
-    def default_format(self, evaluation, form) -> str:
-        value = self.value
-        return '"' + value.__str__() + '"'
-
-    def get_sort_key(self, pattern_sort=False):
-        if pattern_sort:
-            return super().get_sort_key(True)
-        else:
-            return [0, 1, self.value, 0, 1]
-
-    def sameQ(self, other) -> bool:
-        """Mathics SameQ"""
-        # FIX: check
-        if isinstance(other, ByteArrayAtom):
-            return self.value == other.value
-        return False
-
-    def get_string_value(self) -> str:
-        try:
-            return self.value.decode("utf-8")
-        except:
-            return None
-
-    def to_sympy(self, **kwargs):
-        return None
-
-    def to_python(self, *args, **kwargs) -> str:
-        return self.value
-
-    def __hash__(self):
-        return hash(("ByteArrayAtom", self.value))
-
-    def user_hash(self, update):
-        # hashing a String is the one case where the user gets the untampered
-        # hash value of the string's text. this corresponds to MMA behavior.
-        update(self.value)
-
-    def __getnewargs__(self):
-        return (self.value,)
-
-
-class StringFromPython(String):
-    def __new__(cls, value):
-        self = super().__new__(cls, value)
-        if isinstance(value, sympy.NumberSymbol):
-            self.value = "sympy." + str(value)
-
-        # Note that the test is done with math.inf first.
-        # This is to use float's ==, which may not strictly be necessary.
-        if math.inf == value:
-            self.value = "math.inf"
-        return self
 
 
 def get_default_value(name, evaluation, k=None, n=None):
