+++ mathics/core/systemsymbols.py
@@ -1,0 +1,25 @@
+# cython: language_level=3
+# -*- coding: utf-8 -*-
+
+from mathics.core.symbols import Symbol
+
+# Some common Symbols. This list is sorted in alpabetic order.
+SymbolAborted = Symbol("$Aborted")
+SymbolAssociation = Symbol("Association")
+SymbolByteArray = Symbol("ByteArray")
+SymbolComplexInfinity = Symbol("ComplexInfinity")
+SymbolDirectedInfinity = Symbol("DirectedInfinity")
+SymbolFailed = Symbol("$Failed")
+SymbolFalse = Symbol("False")
+SymbolInfinity = Symbol("Infinity")
+SymbolList = Symbol("List")
+SymbolMachinePrecision = Symbol("MachinePrecision")
+SymbolMakeBoxes = Symbol("MakeBoxes")
+SymbolN = Symbol("N")
+SymbolNull = Symbol("Null")
+SymbolRule = Symbol("Rule")
+SymbolSequence = Symbol("Sequence")
+SymbolTrue = Symbol("True")
+SymbolUndefined = Symbol("Undefined")
+SymbolLess = Symbol("Less")
+SymbolGreater = Symbol("Greater")
